Fix [sc-1842]: rename widget
Renaming a widget is not allowed byt the backend, so we have to delete and re-create it with the new id.
