+++ apps/dashboard/src/app/widgets/edit/edit-widget.component.scss
@@ -1,4 +1,4 @@
-@import 'variables';
+@import 'apps/dashboard/src/variables';
 
 .edit-widget {
   width: 770px;
